AKAO (PS1): Fix incorrect instrument tuning with Chrono Cross: Dragon God PSF (#124)
This problem occurs because the chCorrection of the shdr chunk on SoundFont 2 cannot represent a fine tune value that exceeds the range of single byte.

To avoid this restriction, the implementation is changed to normalize a fine tune of 100 cents or more. Changes above a semitone are now applied to the unity key.