[NinSnes] Don't increment the timer after EVENT_END ($00 VCMD)
Had to create a new track variable that will skip incrementing the timer when
set to true. The timer is incremented when all tracks are done processing, which
EVENT_END forces to happen. However, the timer is then incremented, and thus
incurs a one tick delay. This normally does not happen (Yoshikazu Yao's later
builds, not yet supported, is a possible exception because there is some sort of
two-pass system in there, but it could also be caused by lag, which vgmtrans
does not need to emulate), as I recall that the next section is then immediately
processed in at least some of the cases.

This commit closes #155.